labelled notes fixes
